rename wcc to bnscli and update
